+++ clang-tools-extra/clang-tidy/bsl/AutoTypeUsageCheck.cpp
@@ -1,0 +1,40 @@
+//===--- AutoTypeUsageCheck.cpp - clang-tidy ------------------------------===//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
+#include "AutoTypeUsageCheck.h"
+#include "clang/AST/ASTContext.h"
+#include "clang/ASTMatchers/ASTMatchFinder.h"
+
+using namespace clang::ast_matchers;
+
+namespace clang {
+namespace tidy {
+namespace bsl {
+
+void AutoTypeUsageCheck::registerMatchers(MatchFinder *Finder) {
+  // var: is return value and has type of function
+  // cxxCtorInitializer(unless(isFundamentalType()))
+  // lambdaExpr param
+  // functionDecl(hasTrailingReturn())
+  Finder->addMatcher(functionDecl().bind("x"), this);
+}
+
+void AutoTypeUsageCheck::check(const MatchFinder::MatchResult &Result) {
+  // FIXME: Add callback implementation.
+  const auto *MatchedDecl = Result.Nodes.getNodeAs<FunctionDecl>("x");
+  if (MatchedDecl->getName().startswith("awesome_"))
+    return;
+  diag(MatchedDecl->getLocation(), "function %0 is insufficiently awesome")
+      << MatchedDecl;
+  diag(MatchedDecl->getLocation(), "insert 'awesome'", DiagnosticIDs::Note)
+      << FixItHint::CreateInsertion(MatchedDecl->getLocation(), "awesome_");
+}
+
+} // namespace bsl
+} // namespace tidy
+} // namespace clang