+++ clang-tools-extra/clang-tidy/bsl/BslTidyModule.cpp
@@ -9,6 +9,7 @@
 #include "../ClangTidy.h"
 #include "../ClangTidyModule.h"
 #include "../ClangTidyModuleRegistry.h"
+#include "AutoTypeUsageCheck.h"
 #include "ClassBaseCheck.h"
 #include "ClassFinalFunctionCheck.h"
 #include "ClassMemberInitCheck.h"
@@ -55,6 +56,8 @@ namespace bsl {
 class BslModule : public ClangTidyModule {
 public:
   void addCheckFactories(ClangTidyCheckFactories &CheckFactories) override {
+    CheckFactories.registerCheck<AutoTypeUsageCheck>(
+        "bsl-auto-type-usage");
     CheckFactories.registerCheck<ClassBaseCheck>(
         "bsl-class-base");
     CheckFactories.registerCheck<ClassFinalFunctionCheck>(